check_tasks/requirements.txt
celery==5.2.7
python-dotenv==1.0.0
psycopg2-binary==2.9.7
redis==4.5.5
importlib-metadata==6.8.0
